Refactor the indexer demo code
