+++ frontend/src/app/home/Pet/Summary.tsx
@@ -25,7 +25,7 @@ export function Summary({ pet }: SummaryProps) {
   const typedText = useTypingEffect(text);
 
   return (
-    <div className="nes-container is-dark with-title h-[240px]">
+    <div className="nes-container is-dark with-title h-[160px]">
       <p className="title">Summary</p>
       <p>{typedText}</p>
     </div>
