+++ frontend/src/app/home/Pet/index.tsx
@@ -5,7 +5,7 @@ import { Actions, PetAction } from "./Actions";
 import { PetDetails } from "./Details";
 import { PetImage, bodies, ears, faces } from "./Image";
 import { Summary } from "./Summary";
-import { Collection } from "./Collection";
+import { AptogotchiCollection } from "@/components/AptogotchiCollection";
 
 export interface Pet {
   name: string;
@@ -22,30 +22,32 @@ export function Pet({ pet, setPet }: PetProps) {
   const [selectedAction, setSelectedAction] = useState<PetAction>("play");
 
   return (
-    <div className="flex flex-row self-center gap-12 m-8">
-      <div className="flex flex-col gap-4 w-[360px]">
-        <PetImage
-          pet={pet}
-          selectedAction={selectedAction}
-          petParts={{
-            body: bodies[pet.parts[0]],
-            ears: ears[pet.parts[1]],
-            face: faces[pet.parts[2]],
-          }}
-          avatarStyle
-        />
-        <PetDetails pet={pet} setPet={setPet} />
-        <Collection />
+    <div className="flex flex-col self-center m-10">
+      <div className="flex flex-row self-center gap-12">
+        <div className="flex flex-col gap-4 w-[360px]">
+          <PetImage
+            pet={pet}
+            selectedAction={selectedAction}
+            petParts={{
+              body: bodies[pet.parts[0]],
+              ears: ears[pet.parts[1]],
+              face: faces[pet.parts[2]],
+            }}
+            avatarStyle
+          />
+          <PetDetails pet={pet} setPet={setPet} />
+        </div>
+        <div className="flex flex-col gap-8 w-[680px] h-full">
+          <Actions
+            selectedAction={selectedAction}
+            setSelectedAction={setSelectedAction}
+            setPet={setPet}
+            pet={pet}
+          />
+          <Summary pet={pet} />
+        </div>
       </div>
-      <div className="flex flex-col gap-8 w-[680px] h-full">
-        <Actions
-          selectedAction={selectedAction}
-          setSelectedAction={setSelectedAction}
-          setPet={setPet}
-          pet={pet}
-        />
-        <Summary pet={pet} />
-      </div>
+      <AptogotchiCollection />
     </div>
   );
 }
